HW7-2: Split the project into components
