+++ hw5/db.json
@@ -22,28 +22,49 @@
       "isCompleted": true
     },
     {
-      "id": 11,
-      "title": "dsd",
-      "deadline": "2023-04-23",
-      "tag": "other",
-      "isCompleted": false
-    },
-    {
-      "id": 12,
-      "title": "dad",
-      "deadline": "2023-04-23",
-      "tag": "health",
-      "isCompleted": false
-    },
-    {
       "id": 13,
       "title": "dad",
       "deadline": "2023-04-23",
       "tag": "other",
+      "isCompleted": true
+    },
+    {
+      "id": 14,
+      "title": "dad",
+      "deadline": "2023-04-24",
+      "tag": "other",
+      "isCompleted": true
+    },
+    {
+      "id": 15,
+      "title": "тмт",
+      "deadline": "2023-04-24",
+      "tag": "other",
+      "isCompleted": true
+    },
+    {
+      "id": 17,
+      "title": "gffdgd",
+      "deadline": "2023-04-25",
+      "tag": "work",
+      "isCompleted": false
+    },
+    {
+      "id": 18,
+      "title": "uihgjgh",
+      "deadline": "2023-04-25",
+      "tag": "other",
+      "isCompleted": false
+    },
+    {
+      "id": 19,
+      "title": "dsd",
+      "deadline": "2023-04-24",
+      "tag": "home",
       "isCompleted": false
     }
   ],
   "counter": {
-    "count": 13
+    "count": 19
   }
 }
