+++ hw5/main.css
@@ -39,298 +39,11 @@ h1 {
 .display-none{
     display: none !important;
 }
-/* Style for the modal */
-.modal {
-    position: fixed;
-    top: 50%;
-    left: 50%;
-    transform: translate(-50%, -50%);
-    z-index: 9999;
-    width: 41vw;
-    height: 45vh;
-}
-.modal button{
-    cursor: pointer;
-    display: flex;
-    align-items: center;
-}
-/* Style for the modal content */
-.modal-content {
-    background-color: #fefefe;
-    margin: auto;
-    padding: 20px;
-    border: 1px solid #888;
 
-    border-radius: 10px;
-}
-
-/* Style for the close button */
-.close {
-    color: #aaa;
-    float: right;
-    font-size: 28px;
-    font-weight: bold;
-    cursor: pointer;
-}
-
-.close:hover,
-.close:focus {
-    color: black;
-    text-decoration: none;
-    cursor: pointer;
-}
-
-/* Style for the form */
-form {
-    display: flex;
-    flex-direction: column;
-}
-
-/* Style for the tags container */
-.modal__tags {
-    display: flex;
-    align-items: center
-}
-
-/* Style for the tag button */
-.tag {
-    border-radius: 6px;
-    padding: 2px 8px;
-    font-weight: 500;
-    font-size: 12px;
-    margin-right: 12px;
-    height: 19px;
-    border: 1px solid white;
-    display: inline;
-    background: #FFECC7;
-    color: #EA8C00;
-}
-.completed .tag{
-    background: #F5F5F5;
-    color: #838383;
-}
-
-.tag-health{
-    color: #0053CF;
-    background: rgba(60, 134, 244, 0.31);
-}
-.tag-work{
-    color: #9747FF;
-    background: #E8D7FF;
-}
-.tag-home{
-    background: #E2F7E2;
-    color: #639462;
-}
-.tag-other{
-    background: #FFECC7;
-    color: #EA8C00;
-}
-.header__main{
-    display: flex;
-    align-items: center;
-}
-.header__title{
-    font-weight: 700;
-    font-size: 32px;
-    line-height: 36px;
-}
-.header__top{
-    justify-content: space-between;
-}
-.header__weather{
-    align-items: center;
-}
-
-.header__weather-temperature{
-    font-weight: 700;
-    font-size: 18px;
-    line-height: 36px;
-    width: 64px;
-    margin: 0 82px 0 24px;
-}
-.header__city{
-    font-weight: 500;
-    font-size: 18px;
-    line-height: 36px;
-}
-.search-input{
-    padding: 16px 21px;
-
-    width: 100%;
-    height: 52px;
-
-    background: #F5F5F5;
-    border: 1px solid #D2D2D2;
-    border-radius: 12px;
-    font-weight: 500;
-    font-size: 16px;
-    line-height: 20px;
-    color: #000;
-    margin-right: 40px;
-}
-.search-input::placeholder{
-    color: #838383;
-}
-.modal__task-title{
-    box-sizing: border-box;
-    margin-bottom: 18px;
-}
-.header__new-task{
-    padding: 16px 32px;
-    white-space: nowrap;
-    height: 52px;
-    border: none;
-    background: rgba(60, 134, 244, 0.15);
-    border-radius: 12px;
-    font-weight: 800;
-    font-size: 16px;
-    line-height: 20px;
-    cursor: pointer;
-    color: #0053CF;
-
-}
 .row{
     display: flex;
 }
 .column{
     display: flex;
     flex-direction: column;
-}
-.modal__row1{
-    justify-content: space-between;
-    margin-bottom: 92px;
-}
-.modal__row2 {
-    justify-content: space-between;
-}
-.modal__row2 button{
-    width: 49%;
-
-}
-.modal__add-task{
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: flex-end;
-    padding: 16px 32px;
-    gap: 10px;
-
-    height: 52px;
-
-    background: #D3D3D3;
-    border-radius: 12px;
-    font-style: normal;
-    font-weight: 800;
-    font-size: 16px;
-    line-height: 20px;
-    /* identical to box height, or 125% */
-
-    display: flex;
-    align-items: center;
-
-    color: #FFFFFF;
-}
-.modal__add-task--complete{
-    background: #3C86F4;
-    cursor:pointer;
-}
-.modal__cancel{
-    display: flex;
-    flex-direction: row;
-    justify-content: center;
-    align-items: center;
-    padding: 16px 32px;
-    gap: 10px;
-    cursor: pointer;
-    height: 52px;
-
-    border-radius: 12px;
-    font-style: normal;
-    font-weight: 800;
-    font-size: 16px;
-    line-height: 20px;
-
-    color: #0053CF;
-
-}
-.modal__deadline{
-    box-sizing: border-box;
-
-    display: flex;
-    justify-content: center;
-    align-items: center;
-    padding: 10px 16px;
-    gap: 10px;
-
-    height: 40px;
-
-    background: #FFFFFF;
-    border: 2px solid rgba(60, 134, 244, 0.6);
-    border-radius: 12px;
-    font-style: normal;
-    font-weight: 500;
-    font-size: 16px;
-    line-height: 20px;
-    cursor: pointer;
-    color: #1D1D1D;
-}
-.current, .completed{
-    align-items: center;
-    margin-bottom: 29px;
-}
-.current__main-text, .completed__main-text{
-    width: 100%;
-}
-.current__delete{
-    position: absolute;
-    left: 50%;
-    width: 13px;
-    height: 14px;
-    cursor: pointer;
-    background: url(svg/trash_default.svg) no-repeat center center;
-    background-size: 13px 14px;
-}
-.current__delete:hover{
-    background-image: url(svg/trash_hover.svg);
-}
-.current__delete:active{
-    background-image: url(svg/trash_active.svg);
-}
-.current__checkbox, .completed__checkbox{
-    width: 21px;
-    height: 21px;
-    background: #FFFFFF;
-    border: 2px solid #D3D3D3;
-    border-radius: 4px;
-    margin-right: 20px;
-}
-.current__checkbox{
-    cursor: pointer;
-    background: url(svg/checkbox_default.svg) no-repeat center center;
-    background-size: 21px 21px;
-}
-.current__checkbox:hover{
-    background-image: url(svg/checkbox_checked.svg);
-}
-.completed__checkbox{
-    background: url(svg/checkbox_completed.svg) no-repeat center center;
-    background-size: 21px 21px;
-}
-.current__deadline, .completed__deadline{
-    display: flex;
-    align-self: center;
-    font-weight: 500;
-    font-size: 12px;
-    line-height: 16px;
-
-    color: #838383;
-}
-.current__title, .completed__title{
-    margin: 0 0 9px 0;
-    display: flex;
-    align-items: flex-start;
-}
-.completed__title{
-    color:#838383;
 }
