+++ hw5/src/components/Header/Header.css
@@ -1,0 +1,61 @@
+
+.header__main{
+    display: flex;
+    align-items: center;
+}
+.header__title{
+    font-weight: 700;
+    font-size: 32px;
+    line-height: 36px;
+}
+.header__new-task{
+    padding: 16px 32px;
+    white-space: nowrap;
+    height: 52px;
+    border: none;
+    background: rgba(60, 134, 244, 0.15);
+    border-radius: 12px;
+    font-weight: 800;
+    font-size: 16px;
+    line-height: 20px;
+    cursor: pointer;
+    color: #0053CF;
+
+}
+.header__top{
+    justify-content: space-between;
+}
+.header__weather{
+    align-items: center;
+}
+
+.header__weather-temperature{
+    font-weight: 700;
+    font-size: 18px;
+    line-height: 36px;
+    width: 64px;
+    margin: 0 82px 0 24px;
+}
+.header__city{
+    font-weight: 500;
+    font-size: 18px;
+    line-height: 36px;
+}
+.search-input{
+    padding: 16px 21px;
+
+    width: 100%;
+    height: 52px;
+
+    background: #F5F5F5;
+    border: 1px solid #D2D2D2;
+    border-radius: 12px;
+    font-weight: 500;
+    font-size: 16px;
+    line-height: 20px;
+    color: #000;
+    margin-right: 40px;
+}
+.search-input::placeholder{
+    color: #838383;
+}
