+++ hw5/src/components/Modal/Modal.css
@@ -1,0 +1,142 @@
+.modal {
+    position: fixed;
+    top: 50%;
+    left: 50%;
+    transform: translate(-50%, -50%);
+    z-index: 9999;
+    width: 41vw;
+    height: 45vh;
+}
+.modal button{
+    cursor: pointer;
+    display: flex;
+    align-items: center;
+}
+/* Style for the modal content */
+.modal-content {
+    background-color: #fefefe;
+    margin: auto;
+    padding: 20px;
+    border: 1px solid #888;
+
+    border-radius: 10px;
+}
+.modal__row1{
+    justify-content: space-between;
+    margin-bottom: 92px;
+}
+.modal__row2 {
+    justify-content: space-between;
+}
+.modal__row2 button{
+    width: 49%;
+
+}
+.modal__add-task{
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: flex-end;
+    padding: 16px 32px;
+    gap: 10px;
+
+    height: 52px;
+
+    background: #D3D3D3;
+    border-radius: 12px;
+    font-style: normal;
+    font-weight: 800;
+    font-size: 16px;
+    line-height: 20px;
+    /* identical to box height, or 125% */
+
+    display: flex;
+    align-items: center;
+
+    color: #FFFFFF;
+}
+.modal__add-task--complete{
+    background: #3C86F4;
+    cursor:pointer;
+}
+.modal__cancel{
+    display: flex;
+    flex-direction: row;
+    justify-content: center;
+    align-items: center;
+    padding: 16px 32px;
+    gap: 10px;
+    cursor: pointer;
+    height: 52px;
+
+    border-radius: 12px;
+    font-style: normal;
+    font-weight: 800;
+    font-size: 16px;
+    line-height: 20px;
+
+    color: #0053CF;
+
+}
+.modal__deadline{
+    box-sizing: border-box;
+
+    display: flex;
+    justify-content: center;
+    align-items: center;
+    padding: 10px 16px;
+    gap: 10px;
+
+    height: 40px;
+
+    background: #FFFFFF;
+    border: 2px solid rgba(60, 134, 244, 0.6);
+    border-radius: 12px;
+    font-style: normal;
+    font-weight: 500;
+    font-size: 16px;
+    line-height: 20px;
+    cursor: pointer;
+    color: #1D1D1D;
+}
+.modal__task-title{
+    box-sizing: border-box;
+    margin-bottom: 18px;
+}
+/* Style for the tags container */
+.modal__tags {
+    display: flex;
+    align-items: center
+}
+
+/* Style for the tag button */
+.tag {
+    border-radius: 6px;
+    padding: 2px 8px;
+    font-weight: 500;
+    font-size: 12px;
+    margin-right: 12px;
+    height: 19px;
+    border: 1px solid white;
+    display: inline;
+    background: #FFECC7;
+    color: #EA8C00;
+}
+
+
+.tag-health{
+    color: #0053CF;
+    background: rgba(60, 134, 244, 0.31);
+}
+.tag-work{
+    color: #9747FF;
+    background: #E8D7FF;
+}
+.tag-home{
+    background: #E2F7E2;
+    color: #639462;
+}
+.tag-other{
+    background: #FFECC7;
+    color: #EA8C00;
+}
